requirements.txt
# ETL Data Warehouse Pipeline - Python Dependencies
# Install with: pip install -r requirements.txt

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Environment Management
python-dotenv>=1.0.0

# Optional: For dashboard/visualization
# streamlit>=1.28.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
